feat(onboarding): configure models, runtime image, and lane concurrency (#2052)
* fix(images): keep durable assignment success

* feat(onboarding): configure models and runtime image

* feat(onboarding): require models and runtime environment

* fix(onboarding): reuse searchable model picker

* fix(onboarding): make model picker rows readable

* fix(onboarding): stop model search dropping keys

* fix(dev): restore stable macOS server builds

* feat(onboarding): configure parallel agents per lane

* fix(coordinator): allow platform-dependent statvfs casts under clippy

* fix(mcp): project lane_max_sessions fields as i64 in tool schemas

The u32 lane fields emitted nonstandard format=uint32, tripping the MCP
schema contract test. Use the established schemars(with = "i64") pattern
and regenerate the corpus fixture and insta snapshot; the generated UI
types are unchanged.

---------

+++ server/crates/djinn-coordinator/src/cargo_warm_base_gc.rs
@@ -1002,6 +1002,8 @@ impl WarmJobGuard for WarmJobListerGuard {
 
 pub struct StatvfsFreeSpaceGuard;
 impl FreeSpaceGuard for StatvfsFreeSpaceGuard {
+    // statvfs field widths vary by platform (u32 on macOS, u64 on Linux).
+    #[allow(clippy::unnecessary_cast)]
     fn free_space_bytes(&self, path: &Path) -> Result<u64, String> {
         let path = std::ffi::CString::new(path.as_os_str().as_encoded_bytes())
             .map_err(|error| error.to_string())?;
@@ -1011,12 +1013,14 @@ impl FreeSpaceGuard for StatvfsFreeSpaceGuard {
             return Err(std::io::Error::last_os_error().to_string());
         }
         let stat = unsafe { stat.assume_init() };
-        Ok(stat.f_bavail.saturating_mul(stat.f_frsize))
+        Ok((stat.f_bavail as u64).saturating_mul(stat.f_frsize as u64))
     }
 }
 
 pub struct StatvfsFilesystemCapacity;
 impl FilesystemCapacity for StatvfsFilesystemCapacity {
+    // statvfs field widths vary by platform (u32 on macOS, u64 on Linux).
+    #[allow(clippy::unnecessary_cast)]
     fn capacity(&self, path: &Path) -> Result<CapacitySnapshot, String> {
         let path = std::ffi::CString::new(path.as_os_str().as_encoded_bytes())
             .map_err(|error| error.to_string())?;
@@ -1026,8 +1030,8 @@ impl FilesystemCapacity for StatvfsFilesystemCapacity {
             return Err(std::io::Error::last_os_error().to_string());
         }
         let stat = unsafe { stat.assume_init() };
-        let total_bytes = stat.f_blocks.saturating_mul(stat.f_frsize);
-        let available_bytes = stat.f_bavail.saturating_mul(stat.f_frsize);
+        let total_bytes = (stat.f_blocks as u64).saturating_mul(stat.f_frsize as u64);
+        let available_bytes = (stat.f_bavail as u64).saturating_mul(stat.f_frsize as u64);
         Ok(CapacitySnapshot {
             total_bytes,
             available_bytes,
